+++ state/Vehicle.go
@@ -3,8 +3,8 @@ package state
 type Vehicle struct {
 	X             int
 	Y             int
-	Velocity      float32
-	Angle         float32
+	Velocity      float64
+	Angle         float64
 	Endurance     int
 	Team_id       int
 	Max_health    int
